Skills and Items
I made one skill, or rather call its animation according to the conditions of the presence of the skill in the skill slot and the presence of the necessary weapon in the hands. Also today I reconfigured the items and made it so that when an item is inserted into a profile slot, the item will also be visually put on the character.